Refactored configuration parsing/validation

+++ core/hitcounter/direction.go
@@ -2,7 +2,7 @@ package hitcounter
 
 type Direction interface {
 	CleanUp(clock int32)
-	CleanUpTime() int32
+	CleanUpTime() float64
 	Hit(clock int32, val interface{}) bool
 	Name() string
 }
